Move invoke tasks and utils
So they don't have to have wonky names
