+++ tasks/__init__.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+"""Consolidates invoke tasks and utilities
+
+"""
+from .tasks import *  # noqa
+# from .utils import *  # noqa (might not need to include)
